requirements.txt
# Standard libraries
numpy
pandas
matplotlib
seaborn
pillow

# TensorFlow and Keras
tensorflow

# Streamlit
streamlit
